+++ fl_jpush/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 4.1.0
+
+* android添加消息删除回调支持,`onNotifyMessageDismiss`
+* 区分通知和自定义消息回调 `JPushMessage`、`JPushNotificationMessage`,
+*
+
 ## 4.0.2
 
 * 修改`NotificationSettingsWithIOS`的部分默认值为`false`